--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/transformer/QueryTransformer.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/transformer/QueryTransformer.java
@@ -12,7 +12,6 @@ import com.starrocks.analysis.FunctionCallExpr;
 import com.starrocks.analysis.LimitElement;
 import com.starrocks.analysis.OrderByElement;
 import com.starrocks.analysis.SlotRef;
-import com.starrocks.catalog.FunctionSet;
 import com.starrocks.catalog.Type;
 import com.starrocks.common.TreeNode;
 import com.starrocks.qe.ConnectContext;
@@ -275,24 +274,6 @@ class QueryTransformer {
                 WindowTransformer.reorderWindowOperator(windowOperators, columnRefFactory, subOpt);
         for (WindowTransformer.PartitionGroup partitionGroup : partitionGroups) {
             for (WindowTransformer.SortGroup sortGroup : partitionGroup.getSortGroups()) {
-                // Put rank classic window functions of same sort group to the end
-                sortGroup.getWindowOperators().sort((w1, w2) -> {
-                    List<CallOperator> leftCallOperators = Lists.newArrayList(w1.getWindowCall().values());
-                    List<CallOperator> rightCallOperators = Lists.newArrayList(w2.getWindowCall().values());
-                    boolean isLeftRank = leftCallOperators.size() == 1 &&
-                            FunctionSet.ROW_NUMBER.equalsIgnoreCase(leftCallOperators.get(0).getFnName());
-                    boolean isRightRank = rightCallOperators.size() == 1 &&
-                            FunctionSet.ROW_NUMBER.equalsIgnoreCase(rightCallOperators.get(0).getFnName());
-                    if (isLeftRank && isRightRank) {
-                        return 0;
-                    } else if (isLeftRank) {
-                        return 1;
-                    } else if (isRightRank) {
-                        return -1;
-                    } else {
-                        return 0;
-                    }
-                });
                 for (LogicalWindowOperator logicalWindowOperator : sortGroup.getWindowOperators()) {
                     LogicalWindowOperator newLogicalWindowOperator =
                             new LogicalWindowOperator.Builder().withOperator(logicalWindowOperator)