## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #7739

## Enhancement

```sql
-- Q1
select * from (
    select *, rank() over (order by v3) as rk from t0
) sub_t0
order by rk limit 5;

-- Q2
select * from (
    select *, rank() over (partition by v2 order by v3) as rk from t0
) sub_t0
order by rk limit 2, 5;
```

**The optimization is similar to the previous work, by introducing a component `partitioner` to filter data. And I won't go into details about the implementation of `partitioner` here, you can refer to #5885 for design details**

**We will add a transform rule named `PushDownLimitRankingWindowRule` to do the optimitation**

* For query Q1, we can change the `sort operator(order by v3)` below the `window operator` to `topn operator(order by v3, limit 5)`
* For query Q2, we can add an extra `partition topn operator(order by v2,v3, limit 7)` before `window operator`

**And, it should be emphasized that the above optimization can be applied if and only if the following conditions are met**

* no other window function share the same order group with the ranking window function, because we cannot eliminate the full sort, so there is no need to do this optimization( may be lead to negative gain )
    ```sql
    -- Counter example
    select * from (
        select *, 
        rank() over (order by v3) as rk,
        sum(v1) over (order by v3) as sum_v1
        from t0
    ) sub_t0
    order by rk limit 5;
    ```
* the ranking column must be the only one or the first column of the order by columns
    ```sql
    -- Counter example
    select * from (
        select *, rank() over (order by v3) as rk from t0
    ) sub_t0
    order by v1, rk limit 5;
    ```
* the ranking order by column must be `asc`
    ```sql
    -- Counter example
    select * from (
        select *, rank() over (order by v3) as rk from t0
    ) sub_t0
    order by rk desc limit 5;
    ```

## Test

1fe/3be(64c 128g)
tpcds-100g

The following sql is modified from Q-67 by changing `where rk < 100` to `order by rk,xxx limit 100`

```sql
select  *
from (select i_category
            ,i_class
            ,i_brand
            ,i_product_name
            ,d_year
            ,d_qoy
            ,d_moy
            ,s_store_id
            ,sumsales
            ,rank() over (partition by i_category order by sumsales desc) rk
      from (select i_category
                  ,i_class
                  ,i_brand
                  ,i_product_name
                  ,d_year
                  ,d_qoy
                  ,d_moy
                  ,s_store_id
                  ,sum(coalesce(ss_sales_price*ss_quantity,0)) sumsales
            from store_sales
                ,date_dim
                ,store
                ,item
       where  ss_sold_date_sk=d_date_sk
          and ss_item_sk=i_item_sk
          and ss_store_sk = s_store_sk
          and d_month_seq between 1200 and 1200+11
       group by  rollup(i_category, i_class, i_brand, i_product_name, d_year, d_qoy, d_moy,s_store_id))dw1) dw2
order by rk, i_category
        ,i_class
        ,i_brand
        ,i_product_name
        ,d_year
        ,d_qoy
        ,d_moy
        ,s_store_id
        ,sumsales
        ,rk
limit 100;
```


| Before | After |
|:--|:--|
| 20s | 7.8s |