@wanpengfei-git 2022-06-24
#### [FE PR Coverage check]

:heart_eyes: **pass** : 63 / 66 (95.45%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownLimitRankingWindowRule.java|58|61|95.08%|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateRankingWindowRule.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/RuleType.java|2|2|100.00%|