@wanpengfei-git 2022-06-29
#### [FE PR Coverage check]

:heart_eyes: **pass** : 64 / 67 (95.52%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownLimitRankingWindowRule.java|55|58|94.83%|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateRankingWindowRule.java|5|5|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/RuleType.java|2|2|100.00%|